[pf-437] Encapsulated cat related functionnality into the cat algorithm file
